fix(renderer): only add an overlay to titanium ores if commission location matches
Should help to get rid of overlays on non-titanium blocks (e.g. The Forge, as it uses the same block for decoration).

Definitely not the best workaround as the Titanium Miner commission will still show the overlay on those locations (since it doesn't require Titanium to be mined at X place), but at least it's better than nothing.